Start refactoring of new printpdf API
