+++ src_old/pdf_resources.rs
