+++ src_old/xmp_metadata.rs
